finish 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,11 +1,14 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
+// the following function is still not needed now
 // blogsRouter.get('/:id', (request, response, next) => {
 //   Blog.findById(request.params.id)
 //     .then(blog => {
@@ -18,24 +21,52 @@ blogsRouter.get('/', async (request, response) => {
 //     .catch(error => next(error))
 // })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
   if (!body.author || !body.url) return response.status(400).end()
+
+  // const token = request.token
+  // const decodedToken = jwt.verify(token, process.env.SECRET)
+  // if (!token || !decodedToken.id) {
+  //   return response.status(401).json({ error: 'token missing or invalid' })
+  // }
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  // const token = request.token
+  // const decodedToken = jwt.verify(token, process.env.SECRET)
+  // if(!token || !decodedToken.id) {
+  //   return response.status(401).json({ error: 'token missing or invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  console.log(user)
+  const blogToDelete = await Blog.findById(request.params.id)
+  console.log('blog creater id: ', blogToDelete.user._id.toString())
+
+  console.log('delete requester id', user._id.toString())
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'unauthorized deletion attemp' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
